+++ src/test/java/CreateCourierTest.java
@@ -1,4 +1,5 @@
 import io.qameta.allure.junit4.DisplayName;
+import org.junit.After;
 import ya.sprint3.api.ApiSettings;
 import ya.sprint3.api.HttpStatus;
 import io.restassured.RestAssured;
@@ -13,23 +14,23 @@ import static org.junit.Assert.*;
 
 public class CreateCourierTest {
 
+    Courier courier;
     ApiSettings apiSettings = new ApiSettings();
     CourierSteps courierSteps = new CourierSteps();
 
     @Before
     public void setUp() {
         RestAssured.baseURI = "http://qa-scooter.praktikum-services.ru";
+        courier = courierSteps.createCourierWithRandomParameters();
+        apiSettings.pingServer();
     }
 
     @Test
     @DisplayName("Тест корректности работы запроса создания нового курьера")
     @Description("Тест корректности работы запроса создания нового курьера")
     public void createNewCourier() {
-        apiSettings.pingServer();
-        Courier courier;
-        courier = courierSteps.createCourierWithRandomParameters();
         ValidatableResponse response = courierSteps.createNewCourier(courier);
-        assertEquals(response.extract().statusCode(), HttpStatus.CREATED.getValue());
+        assertEquals(HttpStatus.CREATED.getValue(), response.extract().statusCode());
         assertTrue(response.extract().body().path("ok"));
     }
 
@@ -37,56 +38,52 @@ public class CreateCourierTest {
     @DisplayName("Невозможно создать двух одинаковых курьеров")
     @Description("Невозможно создать двух одинаковых курьеров")
     public void checkWhatDoNotPossibleCreateTwoSameCouriers() {
-        apiSettings.pingServer();
-        Courier courier;
-        courier = courierSteps.createCourierWithRandomParameters();
         ValidatableResponse creatingFirstCourier = courierSteps.createNewCourier(courier);
-        assertEquals(creatingFirstCourier.extract().statusCode(), HttpStatus.CREATED.getValue());
+        assertEquals(HttpStatus.CREATED.getValue(), creatingFirstCourier.extract().statusCode());
         ValidatableResponse creatingSecondCourier = courierSteps.createNewCourier(courier);
-        assertEquals(creatingSecondCourier.extract().statusCode(), HttpStatus.CONFLICT.getValue());
+        assertEquals(HttpStatus.CONFLICT.getValue(), creatingSecondCourier.extract().statusCode());
     }
 
     @Test
     @DisplayName("Проверка невозможности создания курьера без логина")
     @Description("Проверка невозможности создания курьера без логина")
     public void checkPossibleCreateCourierWithoutLogin() {
-        apiSettings.pingServer();
-        Courier courierWithoutLogin;
-        courierWithoutLogin = courierSteps.createCourierWithRandomParameters();
-        courierWithoutLogin.setLogin(null);
-        ValidatableResponse creatingCourierWithoutLogin = courierSteps.createNewCourier(courierWithoutLogin);
-        assertEquals(creatingCourierWithoutLogin.extract().statusCode(), HttpStatus.BAD_REQUEST.getValue());
-        assertEquals(creatingCourierWithoutLogin.extract().body().path("message"), "Недостаточно данных для создания учетной записи");
+        courier.setLogin(null);
+        ValidatableResponse creatingCourierWithoutLogin = courierSteps.createNewCourier(courier);
+        assertEquals(HttpStatus.BAD_REQUEST.getValue(), creatingCourierWithoutLogin.extract().statusCode());
+        assertEquals("Недостаточно данных для создания учетной записи", creatingCourierWithoutLogin.extract().body().path("message"));
     }
 
     @Test
     @DisplayName("Проверка невозможности создания курьера без пароля")
     @Description("Проверка невозможности создания курьера без пароля")
     public void checkPossibleCreateCourierWithoutPassword() {
-        apiSettings.pingServer();
-        Courier courierWithoutPassword;
-        courierWithoutPassword = courierSteps.createCourierWithRandomParameters();
-        courierWithoutPassword.setPassword(null);
-        ValidatableResponse creatingCourierWithoutPassword = courierSteps.createNewCourier(courierWithoutPassword);
-        assertEquals(creatingCourierWithoutPassword.extract().statusCode(), HttpStatus.BAD_REQUEST.getValue());
-        assertEquals(creatingCourierWithoutPassword.extract().body().path("message"), "Недостаточно данных для создания учетной записи");
+        courier.setPassword(null);
+        ValidatableResponse creatingCourierWithoutPassword = courierSteps.createNewCourier(courier);
+        assertEquals(HttpStatus.BAD_REQUEST.getValue(), creatingCourierWithoutPassword.extract().statusCode());
+        assertEquals("Недостаточно данных для создания учетной записи", creatingCourierWithoutPassword.extract().body().path("message"));
     }
 
     @Test
     @DisplayName("Нельзя создать курьера с уже используемым логином")
     @Description("Нельзя создать курьера с уже используемым логином")
     public void doNotPossibleCreateCourierWithAlreadyExistingLogin() {
-        apiSettings.pingServer();
-        Courier courier;
-        courier = courierSteps.createCourierWithRandomParameters();
         ValidatableResponse createFirstCourier = courierSteps.createNewCourier(courier);
-        assertEquals(createFirstCourier.extract().statusCode(), HttpStatus.CREATED.getValue());
+        assertEquals(HttpStatus.CREATED.getValue(), createFirstCourier.extract().statusCode());
 
         ValidatableResponse createCourierWithExistLogin = courierSteps.createNewCourier(courier);
         assertEquals(createCourierWithExistLogin.extract().statusCode(), HttpStatus.CONFLICT.getValue());
-        assertEquals(createCourierWithExistLogin.extract().body().path("message"), "Этот логин уже используется.");
+        assertEquals("Этот логин уже используется.", createCourierWithExistLogin.extract().body().path("message"));
+    }
 
-        courierSteps.deleteCourier(courier);
+    @After
+    @DisplayName("Очищаем БД, если были созданы данные")
+    @Description("Очищаем БД, если были созданы данные")
+    public void clearData() {
+        if (courierSteps.checkCourierExists(courier.getLogin(), courier.getPassword())) {
+            courierSteps.deleteCourier(courier);
+        }
+        assertFalse(courierSteps.checkCourierExists(courier.getLogin(), courier.getPassword()));
     }
 
 }
